% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iom_pal.R
\name{iom_pal}
\alias{iom_pal}
\title{IOM Color Palettes}
\usage{
iom_pal(n = NULL, name, ...)
}
\arguments{
\item{n}{Integer. Number of different colors in the palette.
Minimum depends on the palette (2),
maximum depends on the palette (5, 7, or 10).
If \code{NULL}, the maximum number of colors for the palette is used.}

\item{name}{Character. A palette name from the IOM branding.

we have:
\itemize{
\item 2 qualitative palettes - \code{pal_iom} and \code{pal_iom_region}
\item 5 sequential palettes - \code{pal_blue}, \code{pal_un}, \code{pal_green},
\code{pal_red}, \code{pal_yellow}, \code{pal_grey}
\item 2 diverging palette - \code{pal_blue_red} & \code{pal_navy_rev}
}}

\item{...}{Additional arguments passed to \link[grDevices:colorRamp]{grDevices::colorRampPalette}.}
}
\value{
A character vector of hex color codes representing the palette.
}
\description{
Creates a color palette from IOM branding.
}
\details{
This function generates a color palette based on the specified
IOM branding palette name.
If the number of colors \code{n} is not within the valid range for
the palette, warnings are issued and the closest valid number of
colors is used.
}
\examples{
# Display all IOM color palette with filters

# Qualitative Palette:
# When to Use: Use a qualitative palette when you have categorical data,
# meaning data that represents different groups or categories without any
# inherent order.
# Example: Different colors for different countries
iom_pal_display_all(type = "qualitative")
iom_pal(n =3, "pal_iom")

# Sequential Palette:
# When to Use: Use a sequential palette when you have numerical data that
# progresses from low to high. This type of palette shows a gradient of 
# colors, usually from light to dark.
#Example: Temperature changes over a year, population density, or income levels.
iom_pal_display_all(n = 3, type = "sequential")
iom_pal_display_all(n = 5, type = "sequential")
iom_pal(3, "pal_blue")
iom_pal(5, "pal_un")
iom_pal(12, "pal_red")

# Diverging Palette:
# When to Use: Use a diverging palette when you have numerical data with a 
# meaningful midpoint, such as zero, and you want to show variation on both 
# sides of this midpoint.
# Example: Anomalies (deviations from a norm),
# survey responses ranging from negative to positive.
# for diverging palette, you can not get more than 9 different colors
iom_pal_display_all(type = "diverging")
}
