% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iom_scale.R
\name{scale_color_iom_c}
\alias{scale_color_iom_c}
\alias{scale_color_iom_d}
\alias{scale_color_iom_b}
\alias{scale_colour_iom_c}
\alias{scale_colour_iom_d}
\alias{scale_colour_iom_b}
\alias{scale_fill_iom_c}
\alias{scale_fill_iom_d}
\alias{scale_fill_iom_b}
\title{IOM ggplot2 color scales}
\usage{
scale_color_iom_c(
  ...,
  palette = 1,
  direction = 1,
  na.value = "#E9E9E9",
  guide = "colourbar"
)

scale_color_iom_d(
  ...,
  palette = 1,
  direction = 1,
  nmax = NULL,
  order = NULL,
  na.value = "#E9E9E9"
)

scale_color_iom_b(
  ...,
  palette = 1,
  direction = 1,
  nmax = NULL,
  order = NULL,
  na.value = "#E9E9E9",
  guide = "coloursteps"
)

scale_colour_iom_c(
  ...,
  palette = 1,
  direction = 1,
  na.value = "#E9E9E9",
  guide = "colourbar"
)

scale_colour_iom_d(
  ...,
  palette = 1,
  direction = 1,
  nmax = NULL,
  order = NULL,
  na.value = "#E9E9E9"
)

scale_colour_iom_b(
  ...,
  palette = 1,
  direction = 1,
  nmax = NULL,
  order = NULL,
  na.value = "#E9E9E9",
  guide = "coloursteps"
)

scale_fill_iom_c(
  ...,
  palette = 1,
  direction = 1,
  na.value = "#E9E9E9",
  guide = "colourbar"
)

scale_fill_iom_d(
  ...,
  palette = 1,
  direction = 1,
  nmax = NULL,
  order = NULL,
  na.value = "#E9E9E9"
)

scale_fill_iom_b(
  ...,
  palette = 1,
  direction = 1,
  nmax = NULL,
  order = NULL,
  na.value = "#E9E9E9",
  guide = "coloursteps"
)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[=discrete_scale]{discrete_scale()}} or
\code{\link[=continuous_scale]{continuous_scale()}} to control name, limits, breaks, labels and so forth}

\item{palette}{If a string, will use that named palette.
it can be any of the IOM brandied palettes:
\itemize{
\item 2 qualitative palettes - \code{pal_iom} and \code{pal_iom_region}
\item 5 sequential palettes - \code{pal_blue}, \code{pal_navy}, \code{pal_green},
\code{pal_red}, \code{pal_yellow}, \code{pal_grey}
\item 2 diverging palette - \code{pal_blue_red} & \code{pal_navy_rev}
}

If a number, will
index into the list of palettes of appropriate \code{type}}

\item{direction}{Sets the order of colors in the scale. If 1, the default,
colors are as output by \code{\link[=iom_pal]{iom_pal()}}. If -1, the order of colors is reversed}

\item{na.value}{Missing values will be replaced with this value.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{nmax}{Maximum number of different colors the palette should contain.
If not provided, is calculated automatically
from the data.}

\item{order}{Numeric vector listing the order in which the colors should be
used. Default is \code{1:nmax}.}
}
\value{
A discrete, continuous or binned \code{Scale} object
}
\description{
Color scales help differentiate between data points, making it easier to
identify patterns, trends, and outliers.
}
\details{
You will need to select the right type of color scale depending on your data
\itemize{
\item Discrete Color Scales:scale_..._iom_d -  Use these when you have
categorical data  and that require a qualitative color palette.
\item Continuous Color Scales: scale_..._iom_c -Use these for numerical data
that progresses from low to high and that require a sequential color palette.
\item Diverging Color Scales: scale_..._iom_b - Use these when you have data
with a meaningful midpoint and that require a diverging color palette.
}

depending on the geometry use, either fill or color - or a combination
}
\examples{
library(ggplot2)
library(iomthemes)


# Discrete Color Scales:scale_..._iom_d -  Use these when you have 
#     categorical data  and that require a qualitative color palette. 

df <- data.frame(
  date = factor(rep(2010:2020, 2)),
  val = c(
    10.2, 10.2, 10.2, 11.5, 14,
    15.8, 17, 19.5, 20.5, 24.5, 20.6,
    4.9, 4.9, 4.9, 5.5, 5.5, 5.6,
    5.6, 5.7, 5.7, 5.8, 5.9
  ),
  set = rep(c("Set 1", "Set 2"), each = 11)
)
ggplot(
  data = df,
  aes(x = date, y = val, fill = set)
) +
  geom_col(
    position = position_stack(reverse = TRUE),
    width = 0.8
  ) +
  labs(
    title = "My title including 
              <span style='color:#0033A0;'>highlight color</span> | 2010-2020",
    subtitle = "My really subtitle that will cover more than one line and
              that will be automatically wrapped thanks to ggtext so I don't 
              have to manually put a line break",
    y = "values",
    caption = "Note: As it's HTML text the line break symbol is
               not '\\\\n'<br>Source: Dummy data"
  ) +
  scale_y_continuous(
    expand = expansion(c(0, 0.01))
  ) +
  scale_fill_iom_d() +
  theme_iom(
    grid = "Y",
    axis = "X",
    axis_title = "Y"
  )

# Continuous Color Scales: scale_..._iom_c -Use these for numerical data that progresses
#  from low to high and that require a sequential color palette. 

# Diverging Color Scales: scale_..._iom_b - Use these when you have data 
# 'with a meaningful midpoint and that require a diverging color palette. 
}
