% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iomthemes-package.R
\docType{package}
\name{iomthemes-package}
\alias{iomthemes}
\alias{iomthemes-package}
\title{iomthemes: IOM 'ggplot2' Theme and Colour Palettes}
\description{
A 'ggplot2' theme and color palettes following IOM visual branding recommendations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iom/iomthemes}
  \item \url{https://iom.github.io/iomthemes/}
  \item Report bugs at \url{https://github.com/iom/iomthemes/issues}
}

}
\author{
\strong{Maintainer}: Edouard Legoupil \email{elegoupil@iom.int}

Other contributors:
\itemize{
  \item IOM [copyright holder]
}

}
\keyword{internal}
